+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ﻿## 1.0.1 (Feb, 9, 2016)
     - Implemented excel unit tests
     - Implemented SQL unit tests
+    - Implemented Import-XML function and tests
 
 ## 1.0.0 (Feb, 8, 2016)
     - Initial Release